+++ packages/mpv-player/src/index.ts
@@ -1,2 +1,3 @@
 export { MpvPlayerView } from "./MpvPlayerView";
+export { mpvAudio, type MpvAudioProgress } from "./MpvAudio";
 export * from "./MpvPlayer.types";